+++ NEWS.md
@@ -3,7 +3,22 @@ title: "Release notes for `ECCTools`"
 output: html_document
 ---
 
+
+# ECCTools 0.1.3 (14/07/2022)
+
+* exports_to_balancing function and corresponding tests added
+* losses_to_balancing function and corresponding tests added
+* non_energy_uses_to_balancing function and corresponding tests added
+* specify_other_elec_heat_production function and corresponding tests added
+* specify_elect_heat_markets function and corresponding tests modified so that it works with the new specify_other_elec_heat_production function too
+* New specify_losses_as_industry function to treat losses as a transformation industry
+* ECC figure for country A updated
+* All tests passing (332)
+* Test coverage at 100%.
+
+
 # ECCTools 0.1.2 (07/01/2022)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5998974.svg)](https://doi.org/10.5281/zenodo.5998974)
 
 * Some fixes on the specify_elect_heat_fossil_fuels() function
 * Issues with badges and README file solved